toolchain: Simplify platform detection logic
The scripts/build_toolchain.sh file is clutered with repetitive
if statements that detect and set the proper variables for
toolchain retrieval of compilation.

Not only is there redundant duplication of code, it does not work
on Cygwin, whose uname returned string varies by the host version
of Windows.

Switching to case statements simplifies the code and gives it an
overall nicer aesthetic. Additionally wrapping the prebuilt
toolchain retrieval code into a function further simplifies the
switch statements.

Signed-off-by: Aun-Ali Zaidi <admin@kodeit.net>